backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
nibabel
alive_progress